requirements.txt
# LangGraph 核心依赖
langgraph==1.0.4
langchain==1.1.3
langchain-core==1.1.3

# OpenAI 集成
langchain-openai==1.1.1

# 数据验证与类型提示
pydantic==2.12.2
typing-extensions==4.15.0

# 环境变量管理
python-dotenv==1.1.0

# HTTP 请求（用于 KIE API）
requests==2.32.5
supabase==2.24.0

# FastAPI 服务层
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==2.2.1

# LangGraph SDK 客户端
langgraph-sdk==0.1.51